# it/Ingegneria_del_Software/GeoControl/doc/swagger_v1.yaml
openapi: 3.0.0
info:
  title: GeoControl API
  description: |
    **GeoControl System Overview**

    GeoControl is a software system designed for monitoring physical and environmental variables in various contexts: from hydrogeological analyses of mountain areas to the surveillance of historical buildings, and even the control of internal parameters (such as temperature or lighting) in residential or working environments.  
    It was commissioned by the Union of Mountain Communities of the Piedmont region for managing the hydrogeological state of its territories. However, thanks to its modular structure, GeoControl has also been commercialized to different public or private entities requiring continuous monitoring of physical parameters.  
    The system meets high reliability requirements and must ensure that no more than six measurements per year, per sensor, are lost. From an organizational perspective, GeoControl uses a token-based authentication mechanism, with different user roles to regulate access levels to the functionalities.

    ---

    **Hierarchical Structure and Components**

    GeoControl’s hierarchy is organized into three main levels:

    1. **Network**  
       - A logical grouping that acts as a container for multiple gateways (and their associated sensors).  
       - Identified by a unique alphanumeric code (chosen at creation time) and can represent, for example, a monitoring network for a single municipality or an entire building.  
...

...

...

...

    ---

...

...
servers:
...
paths:
...
  /networks/{networkCode}/gateways:
    get:
      summary: Retrieve all gateways of a network
      description: This endpoint can be used by any kind of authenticated user
      tags:
        - Gateways
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A list of gateways
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gateway"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
    post:
      summary: Create a new gateway for a network
      description: |
        This endpoint can be used only by Admin and Operator users
        Each request to create a Gateway is considered valid even if it contains nested Sensors. However, only the Gateway data will be processed. Any nested objects included in the request will be ignored and not created through this operation.
      tags:
        - Gateways
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [macAddress]
              properties:
                macAddress:
                  type: string
                  minLength: 1
                  description: MAC address of the gateway
                  example: 00:11:22:33:44:55
                name:
                  type: string
                  description: Name of the gateway
                  example: Main Gateway
                description:
                  type: string
                  description: Description of the gateway
                  example: Gateway at main site
                sensors:
                  type: array
                  items:
                    $ref: "#/components/schemas/Sensor"
            example:
              macAddress: "94:3F:BE:4C:4A:79"
              name: "GW01"
              description: "on-field aggregation node"
      responses:
        "201":
          description: Gateway created
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "400"
                name: "BadRequest"
                message: "/body/xxx' must have required property 'xxx'"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "403":
          description: Insufficient rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "403"
                name: "InsufficientRightsError"
                message: "Forbidden: Insufficient rights"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "409":
          description: Gateway mac address already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "409"
                name: "ConflictError"
                message: "Entity with code xxxxx already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
  /networks/{networkCode}/gateways/{gatewayMac}:
    get:
      summary: Retrieve a specific gateway
      description: This endpoint can be used by any kind of authenticated user
      tags:
        - Gateways
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
        - name: gatewayMac
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A single gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gateway"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "404":
          description: Network/Gateway not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
    patch:
      summary: Update a gateway
      description: |
        This endpoint can be used only by Admin and Operator users
        Each request to update a Gateway is considered valid even if it contains nestedSensors. However, only the Gateway data will be processed. Any nested objects included in the request will be ignored and not created or updated through this operation. The request must allow the update of all the Gateway fields, including the Gateway mac address.
      tags:
        - Gateways
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
        - name: gatewayMac
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Gateway"
            example:
              macAddress: "1C:6A:66:95:EE:45"
              name: "Updated Name"
              description: "Updated Description"
      responses:
        "204":
          description: Gateway updated
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "400"
                name: "BadRequest"
                message: "/body/xxx' must have required property 'xxx'"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "403":
          description: Insufficient rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "403"
                name: "InsufficientRightsError"
                message: "Forbidden: Insufficient rights"
        "404":
          description: Network/Gateway not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "409":
          description: Gateway mac address already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "409"
                name: "ConflictError"
                message: "Entity with code xxxxx already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
    delete:
      summary: Delete a Gateway
      description: This endpoint can be used only by Admin and Operator users
      tags:
        - Gateways
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
        - name: gatewayMac
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Gateway deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "403":
          description: Insufficient rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "403"
                name: "InsufficientRightsError"
                message: "Forbidden: Insufficient rights"
        "404":
          description: Network/Gateway not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Gateway:
      type: object
      properties:
        macAddress:
          type: string
          minLength: 1
          description: Unique identifier for the gateway (MAC address)
          example: 94:3F:BE:4C:4A:79
        name:
          type: string
          description: Name of the gateway
          example: GW01
        description:
          type: string
          description: Description of the gateway
          example: on-field aggregation node
        sensors:
          type: array
          items:
            $ref: "#/components/schemas/Sensor"
    Sensor:
      type: object
      properties:
        macAddress:
          type: string
          minLength: 1
          description: Unique identifier for the sensor (MAC address)
          example: 71:B1:CE:01:C6:A9
        name:
          type: string
          description: Name of the sensor
          example: TH01
        description:
          type: string
          description: Description of the sensor
          example: External thermometer
        variable:
          type: string
          description: Variable associated with the sensor
          example: temperature
        unit:
          type: string
          description: Unit of the variable
          example: C
...
    ErrorDTO:
      type: object
      required:
        - code
      properties:
        code:
          type: number
          description: HTTP code of the error
          example: 404
        name:
          type: string
          description: Name of the error
          example: NotFoundError
        message:
          type: string
          description: Message describing the type and the cause of the error
          example: Entity not found